+++ Felis.Core/Felis.Core.csproj
@@ -7,10 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.8" />
-    </ItemGroup>
-
-    <ItemGroup>
       <Compile Remove="Models\Client.cs" />
     </ItemGroup>
